refactor: better description
