+++ src/content/features/click-modal-inactive-check.js
